Update models structure & add feed admin + fix celery task detection
This commit simplifies the models structure of feeds. Instead of having model inheritance
we use a single Feed model with an unique url field for the various types of feeds.
Twitter accounts are validated by a regex to test if the entered URL follows the twitter's
account address rule (https://twitter.com/<account name>). This simplifies a lot the admin
interface creation since we don't have to deal with inheritance (that would bring not much
improvement to the architecture nor the logic).

This commit also fix the detection of application tasks by celery. It wasn't working
because of a falty python file that celery couldn't import (crawler.scraping.tasks).